Preliminary progress - Introduction
Browser updated, submitting this to basically have my account show up for you.  (Assuming this even displays on your end... lawl)
The only "change" I made was to the naming of the guide, changing it from "Dynamic Labels" to "Text Labels". I felt this better conveyed that this guide contains info for application to all Labels, both static AND dynamic (FT-equipped labels). A mere suggestion, in the end, as it can go back to what it was, or change to something better suited - I tried to keep it "short", keeping the Nav Menu listing in mind.